First checkpoint successful
Checkpoint: Run your app in Simulator. The text field shouldn’t extend
past the edge of the screen anymore. You should be able to click inside
the text field and enter text using the keyboard (if you’d like, toggle
the software keyboard by pressing Command-K). If you rotate the device
(Command-Left Arrow or Command-Right Arrow) or run the app on a
different device, the text field grows or shrinks to the appropriate
size depending on the device’s orientation and screen size. Note that
the status bar disappears in landscape orientation.