Third checkpoint successful
Test your changes by running Simulator. You can select the text field
and type text into it. When you click the Done button on the keyboard,
the keyboard is dismissed and the label text changes to display the text
in the text field. When you click the Set Default Label Text button, the
label changes from what’s currently displayed in the label to `Default
Text` (the value set by the action you defined earlier).