Fifth checkpoint successful
Run your app. You should be able to click the image view to pull up an
image picker. You’ll need to click OK on the alert that asks for
permission to give the FoodTracker app access to Photos. Then, you can
click the Cancel button to dismiss the picker, or open Camera Roll and
click an image to select it and set it as the image in the image view.